feat(ui): replace unread cap with compact count format
